refactor: command wrapper
